Add files via upload
These .sql files contain the schema for the users and tasks schema. I ran these through pgAdmin4 as tables in which users is the base table and tasks depends on "assigned_user" as a foreign key. There are extraneous bits of this code as it is the raw backup from pgAdmin. Please refer to the schema specifically when running again on another pgAdmin instance.